[Refactor] Move keybinds extension to keybindingService
