+++ src/extensions/core/index.ts
@@ -6,7 +6,6 @@ import './groupNode'
 import './groupNodeManage'
 import './groupOptions'
 import './invertMenuScrolling'
-import './keybinds'
 import './maskeditor'
 import './nodeTemplates'
 import './noteNode'
